Tests API endpoint for 200 status and matching JSON object
