--- api.test.js
@@ -1,24 +1,0 @@
-import React from 'react';
-// import renderer from 'react-test-renderer';
-import { waitFor, screen, cleanup, fireEvent, render} from '@testing-library/react';
-import '@testing-library/jest-dom
-import axiosMock from 'axios';
-
-// import App from './client/components/App';
-
-test('Fake Test', () => { expect(true).toBeTruthy(); });
-
-// describe('App', () => {
-//   test('snapshot renders', () => {
-//     const component = renderer.create(<App />);
-//     let tree = component.toJSON();
-//     expect(tree).toMatchSnapshot();
-//   })
-// })
-
-// test('App renders correctly', () => {
-//   const component = renderer.create(<App />,);
-//   const tree = component.toJSON();
-//   console.log(tree);
-//   expect(tree).toMatchSnapshot();
-// });
