+++ database/index.js
@@ -2,8 +2,7 @@ const mysql = require('mysql');
 
 const db = mysql.createConnection({
   host: 'localhost',
-  user: 'student',
-  password: 'student',
+  user: 'root',
   database: 'gamestart',
 });
