+++ database/seeding.js
@@ -3,8 +3,7 @@ const mysql = require('mysql');
 
 const db = mysql.createConnection({
   host: 'localhost',
-  user: 'student',
-  password: 'student',
+  user: 'root',
   database: 'gamestart',
 });
